fix: fixed simulations
