+++ simulations/vip-356/bscmainnet.ts
@@ -10,7 +10,7 @@ forking(40343666, async () => {
         txResponse,
         [OMNICHAIN_PROPOSAL_SENDER_ABI],
         ["ExecuteRemoteProposal", "StorePayload"],
-        [2, 0],
+        [3, 0],
       );
     },
   });
