+++ simulations/vip-356/bsctestnet.ts
@@ -11,7 +11,7 @@ forking(42106572, async () => {
         txResponse,
         [OMNICHAIN_PROPOSAL_SENDER_ABI],
         ["ExecuteRemoteProposal", "StorePayload"],
-        [2, 0],
+        [3, 0],
       );
     },
   });
