+++ simulations/vip-356/opbnbmainnet.ts
@@ -1,13 +1,13 @@
 import { expectEvents } from "src/utils";
 import { forking, testForkedNetworkVipCommands } from "src/vip-framework";
 
-import vip357 from "../../vips/vip-357/bsctestnet";
+import vip356 from "../../vips/vip-356/bscmainnet";
 import ACCESS_CONTROL_MANAGER_ABI from "./abi/AccessControlManager_ABI.json";
 
-forking(62418214, async () => {
-  testForkedNetworkVipCommands("vip357", await vip357(), {
+forking(28761242, async () => {
+  testForkedNetworkVipCommands("vip357", await vip356(), {
     callbackAfterExecution: async txResponse => {
-      await expectEvents(txResponse, [ACCESS_CONTROL_MANAGER_ABI], ["PermissionRevoked"], [3]);
+      await expectEvents(txResponse, [ACCESS_CONTROL_MANAGER_ABI], ["PermissionRevoked"], [1]);
     },
   });
 });
