+++ simulations/vip-356/opbnbtestnet.ts
@@ -1,13 +1,13 @@
 import { expectEvents } from "src/utils";
 import { forking, testForkedNetworkVipCommands } from "src/vip-framework";
 
-import vip358 from "../../vips/vip-358/bsctestnet";
+import vip356 from "../../vips/vip-356/bsctestnet";
 import ACCESS_CONTROL_MANAGER_ABI from "./abi/AccessControlManager_ABI.json";
 
-forking(6276156, async () => {
-  testForkedNetworkVipCommands("vip358", await vip358(), {
+forking(36325286, async () => {
+  testForkedNetworkVipCommands("vip356", await vip356(), {
     callbackAfterExecution: async txResponse => {
-      await expectEvents(txResponse, [ACCESS_CONTROL_MANAGER_ABI], ["PermissionRevoked"], [11]);
+      await expectEvents(txResponse, [ACCESS_CONTROL_MANAGER_ABI], ["PermissionRevoked"], [1]);
     },
   });
 });
