+++ simulations/vip-357/bsctestnet.ts
@@ -11,7 +11,7 @@ forking(42106572, async () => {
         txResponse,
         [OMNICHAIN_PROPOSAL_SENDER_ABI],
         ["ExecuteRemoteProposal", "StorePayload"],
-        [3, 0],
+        [1, 0],
       );
     },
   });
