--- simulations/vip-357/opbnbtestnet.ts
@@ -1,13 +1,0 @@
-import { expectEvents } from "src/utils";
-import { forking, testForkedNetworkVipCommands } from "src/vip-framework";
-
-import vip357 from "../../vips/vip-357/bsctestnet";
-import ACCESS_CONTROL_MANAGER_ABI from "./abi/AccessControlManager_ABI.json";
-
-forking(36325286, async () => {
-  testForkedNetworkVipCommands("vip357", await vip357(), {
-    callbackAfterExecution: async txResponse => {
-      await expectEvents(txResponse, [ACCESS_CONTROL_MANAGER_ABI], ["PermissionRevoked"], [4]);
-    },
-  });
-});
