--- simulations/vip-358/arbitrumsepolia.ts
@@ -1,13 +1,0 @@
-import { expectEvents } from "src/utils";
-import { forking, testForkedNetworkVipCommands } from "src/vip-framework";
-
-import vip358 from "../../vips/vip-358/bsctestnet";
-import ACCESS_CONTROL_MANAGER_ABI from "./abi/AccessControlManager_ABI.json";
-
-forking(62418214, async () => {
-  testForkedNetworkVipCommands("vip358", await vip358(), {
-    callbackAfterExecution: async txResponse => {
-      await expectEvents(txResponse, [ACCESS_CONTROL_MANAGER_ABI], ["PermissionRevoked"], [11]);
-    },
-  });
-});
