--- simulations/vip-359/ethereum.ts
@@ -1,13 +1,0 @@
-import { expectEvents } from "src/utils";
-import { forking, testForkedNetworkVipCommands } from "src/vip-framework";
-
-import vip359 from "../../vips/vip-359/bscmainnet";
-import ACCESS_CONTROL_MANAGER_ABI from "./abi/AccessControlManager_ABI.json";
-
-forking(20274111, async () => {
-  testForkedNetworkVipCommands("vip359", await vip359(), {
-    callbackAfterExecution: async txResponse => {
-      await expectEvents(txResponse, [ACCESS_CONTROL_MANAGER_ABI], ["PermissionRevoked"], [14]);
-    },
-  });
-});
