+++ vips/vip-356/bsctestnet.ts
@@ -3,9 +3,15 @@ import { NETWORK_ADDRESSES } from "src/networkAddresses";
 import { LzChainId, ProposalType } from "src/types";
 import { makeProposal } from "src/utils";
 
-const { arbitrumsepolia, sepolia } = NETWORK_ADDRESSES;
+const { arbitrumsepolia, sepolia, opbnbtestnet } = NETWORK_ADDRESSES;
 export const ARBITRUM_SEPOLIA_ACM = "0xa36AD96441cB931D8dFEAAaC97D3FaB4B39E590F";
 export const SEPOLIA_ACM = "0xbf705C00578d43B6147ab4eaE04DBBEd1ccCdc96";
+export const OPBNBTESTNET_ACM = "0x049f77F7046266d27C3bC96376f53C17Ef09c986";
+
+export const ARBITRUM_SEPOLIA_PRIME = "0xadb04ac4942683bc41e27d18234c8dc884786e89";
+export const ARBITRUM_SEPOLIA_PLP = "0xe82c2c10f55d3268126c29ec813dc6f086904694";
+export const SEPOLIA_PRIME = "0x2Ec432F123FEbb114e6fbf9f4F14baF0B1F14AbC";
+export const SEPOLIA_PLP = "0x15242a55Ad1842A1aEa09c59cf8366bD2f3CE9B4";
 
 const vip356 = () => {
   const meta = {
@@ -22,7 +28,11 @@ const vip356 = () => {
       {
         target: ARBITRUM_SEPOLIA_ACM,
         signature: "revokeCallPermission(address,string,address)",
-        params: [ethers.constants.AddressZero, "setRewardTokenSpeeds(address[],uint256[],uint256[])", arbitrumsepolia.GUARDIAN],
+        params: [
+          ethers.constants.AddressZero,
+          "setRewardTokenSpeeds(address[],uint256[],uint256[])",
+          arbitrumsepolia.GUARDIAN,
+        ],
         dstChainId: LzChainId.arbitrumsepolia,
       },
       {
@@ -38,7 +48,11 @@ const vip356 = () => {
       {
         target: ARBITRUM_SEPOLIA_ACM,
         signature: "revokeCallPermission(address,string,address)",
-        params: [ethers.constants.AddressZero, "setLastRewardingBlocks(address[],uint32[],uint32[])", arbitrumsepolia.GUARDIAN],
+        params: [
+          ethers.constants.AddressZero,
+          "setLastRewardingBlocks(address[],uint32[],uint32[])",
+          arbitrumsepolia.GUARDIAN,
+        ],
         dstChainId: LzChainId.arbitrumsepolia,
       },
       {
@@ -60,6 +74,175 @@ const vip356 = () => {
         target: SEPOLIA_ACM,
         signature: "revokeCallPermission(address,string,address)",
         params: [ethers.constants.AddressZero, "setLastRewardingBlock(address[],uint32[],uint32[])", sepolia.GUARDIAN],
+        dstChainId: LzChainId.sepolia,
+      },
+
+      {
+        target: ARBITRUM_SEPOLIA_ACM,
+        signature: "revokeCallPermission(address,string,address)",
+        params: [
+          ethers.constants.AddressZero,
+          "updateJumpRateModel(uint256,uint256,uint256,uint256)",
+          arbitrumsepolia.GUARDIAN,
+        ],
+        dstChainId: LzChainId.arbitrumsepolia,
+      },
+      {
+        target: SEPOLIA_ACM,
+        signature: "revokeCallPermission(address,string,address)",
+        params: [
+          ethers.constants.AddressZero,
+          "updateJumpRateModel(uint256,uint256,uint256,uint256)",
+          sepolia.GUARDIAN,
+        ],
+        dstChainId: LzChainId.sepolia,
+      },
+      {
+        target: OPBNBTESTNET_ACM,
+        signature: "revokeCallPermission(address,string,address)",
+        params: [
+          ethers.constants.AddressZero,
+          "updateJumpRateModel(uint256,uint256,uint256,uint256)",
+          opbnbtestnet.GUARDIAN,
+        ],
+        dstChainId: LzChainId.opbnbtestnet,
+      },
+
+      {
+        target: ARBITRUM_SEPOLIA_ACM,
+        signature: "revokeCallPermission(address,string,address)",
+        params: [ARBITRUM_SEPOLIA_PRIME, "setTokensDistributionSpeed(address[],uint256[])", arbitrumsepolia.GUARDIAN],
+        dstChainId: LzChainId.arbitrumsepolia,
+      },
+      {
+        target: ARBITRUM_SEPOLIA_ACM,
+        signature: "revokeCallPermission(address,string,address)",
+        params: [
+          ARBITRUM_SEPOLIA_PRIME,
+          "setMaxTokensDistributionSpeed(address[],uint256[])",
+          arbitrumsepolia.GUARDIAN,
+        ],
+        dstChainId: LzChainId.arbitrumsepolia,
+      },
+      {
+        target: ARBITRUM_SEPOLIA_ACM,
+        signature: "revokeCallPermission(address,string,address)",
+        params: [ARBITRUM_SEPOLIA_PRIME, "setMaxLoopsLimit(uint256)", arbitrumsepolia.GUARDIAN],
+        dstChainId: LzChainId.arbitrumsepolia,
+      },
+      {
+        target: ARBITRUM_SEPOLIA_ACM,
+        signature: "revokeCallPermission(address,string,address)",
+        params: [ARBITRUM_SEPOLIA_PLP, "updateAlpha(uint128,uint128)", arbitrumsepolia.GUARDIAN],
+        dstChainId: LzChainId.arbitrumsepolia,
+      },
+      {
+        target: ARBITRUM_SEPOLIA_ACM,
+        signature: "revokeCallPermission(address,string,address)",
+        params: [ARBITRUM_SEPOLIA_PLP, "updateMultipliers(address,uint256,uint256)", arbitrumsepolia.GUARDIAN],
+        dstChainId: LzChainId.arbitrumsepolia,
+      },
+      {
+        target: ARBITRUM_SEPOLIA_ACM,
+        signature: "revokeCallPermission(address,string,address)",
+        params: [ARBITRUM_SEPOLIA_PLP, "setStakedAt(address[],uint256[])", arbitrumsepolia.GUARDIAN],
+        dstChainId: LzChainId.arbitrumsepolia,
+      },
+      {
+        target: ARBITRUM_SEPOLIA_ACM,
+        signature: "revokeCallPermission(address,string,address)",
+        params: [ARBITRUM_SEPOLIA_PLP, "addMarket(address,address,uint256,uint256)", arbitrumsepolia.GUARDIAN],
+        dstChainId: LzChainId.arbitrumsepolia,
+      },
+      {
+        target: ARBITRUM_SEPOLIA_ACM,
+        signature: "revokeCallPermission(address,string,address)",
+        params: [ARBITRUM_SEPOLIA_PLP, "setLimit(uint256,uint256)", arbitrumsepolia.GUARDIAN],
+        dstChainId: LzChainId.arbitrumsepolia,
+      },
+      {
+        target: ARBITRUM_SEPOLIA_ACM,
+        signature: "revokeCallPermission(address,string,address)",
+        params: [ARBITRUM_SEPOLIA_PLP, "setMaxLoopsLimit(uint256)", arbitrumsepolia.GUARDIAN],
+        dstChainId: LzChainId.arbitrumsepolia,
+      },
+      {
+        target: ARBITRUM_SEPOLIA_ACM,
+        signature: "revokeCallPermission(address,string,address)",
+        params: [ARBITRUM_SEPOLIA_PLP, "issue(bool,address[])", arbitrumsepolia.GUARDIAN],
+        dstChainId: LzChainId.arbitrumsepolia,
+      },
+      {
+        target: ARBITRUM_SEPOLIA_ACM,
+        signature: "revokeCallPermission(address,string,address)",
+        params: [ARBITRUM_SEPOLIA_PLP, "burn(address)", arbitrumsepolia.GUARDIAN],
+        dstChainId: LzChainId.arbitrumsepolia,
+      },
+
+      {
+        target: SEPOLIA_ACM,
+        signature: "revokeCallPermission(address,string,address)",
+        params: [SEPOLIA_PRIME, "setTokensDistributionSpeed(address[],uint256[])", sepolia.GUARDIAN],
+        dstChainId: LzChainId.sepolia,
+      },
+      {
+        target: SEPOLIA_ACM,
+        signature: "revokeCallPermission(address,string,address)",
+        params: [SEPOLIA_PRIME, "setMaxTokensDistributionSpeed(address[],uint256[])", sepolia.GUARDIAN],
+        dstChainId: LzChainId.sepolia,
+      },
+      {
+        target: SEPOLIA_ACM,
+        signature: "revokeCallPermission(address,string,address)",
+        params: [SEPOLIA_PRIME, "setMaxLoopsLimit(uint256)", sepolia.GUARDIAN],
+        dstChainId: LzChainId.sepolia,
+      },
+      {
+        target: SEPOLIA_ACM,
+        signature: "revokeCallPermission(address,string,address)",
+        params: [SEPOLIA_PLP, "updateAlpha(uint128,uint128)", sepolia.GUARDIAN],
+        dstChainId: LzChainId.sepolia,
+      },
+      {
+        target: SEPOLIA_ACM,
+        signature: "revokeCallPermission(address,string,address)",
+        params: [SEPOLIA_PLP, "updateMultipliers(address,uint256,uint256)", sepolia.GUARDIAN],
+        dstChainId: LzChainId.sepolia,
+      },
+      {
+        target: SEPOLIA_ACM,
+        signature: "revokeCallPermission(address,string,address)",
+        params: [SEPOLIA_PLP, "setStakedAt(address[],uint256[])", sepolia.GUARDIAN],
+        dstChainId: LzChainId.sepolia,
+      },
+      {
+        target: SEPOLIA_ACM,
+        signature: "revokeCallPermission(address,string,address)",
+        params: [SEPOLIA_PLP, "addMarket(address,address,uint256,uint256)", sepolia.GUARDIAN],
+        dstChainId: LzChainId.sepolia,
+      },
+      {
+        target: SEPOLIA_ACM,
+        signature: "revokeCallPermission(address,string,address)",
+        params: [SEPOLIA_PLP, "setLimit(uint256,uint256)", sepolia.GUARDIAN],
+        dstChainId: LzChainId.sepolia,
+      },
+      {
+        target: SEPOLIA_ACM,
+        signature: "revokeCallPermission(address,string,address)",
+        params: [SEPOLIA_PLP, "setMaxLoopsLimit(uint256)", sepolia.GUARDIAN],
+        dstChainId: LzChainId.sepolia,
+      },
+      {
+        target: SEPOLIA_ACM,
+        signature: "revokeCallPermission(address,string,address)",
+        params: [SEPOLIA_PLP, "issue(bool,address[])", sepolia.GUARDIAN],
+        dstChainId: LzChainId.sepolia,
+      },
+      {
+        target: SEPOLIA_ACM,
+        signature: "revokeCallPermission(address,string,address)",
+        params: [SEPOLIA_PLP, "burn(address)", sepolia.GUARDIAN],
         dstChainId: LzChainId.sepolia,
       },
     ],
